Calculator finally completed!

+++ Udemy/day10/calculator_project.py
@@ -19,13 +19,13 @@ operations = {
 "/" : divide
 }
 
-num1 = int(input("What's the first number?: \n")) 
+num1 = float(input("What's the first number?: \n")) 
 
 for symbol in operations:
     print(symbol)
 function = input("Pick an operation from the lines above: \n")
 
-num2 = int(input("What's the second number?: \n")) 
+num2 = float(input("What's the second number?: \n")) 
 
 for symbol in operations:
     if function == symbol:
@@ -38,7 +38,7 @@ keep_going = input(f"Type 'y' to continue calculating with {answer}, or type 'n'
 while keep_going == 'y' or keep_going == 'n':
 
     function = input("Pick an operation: \n")
-    num3 = int(input("What's the next number?: \n"))
+    num3 = float(input("What's the next number?: \n"))
 
     total = answer
 
@@ -48,9 +48,9 @@ while keep_going == 'y' or keep_going == 'n':
     keep_going = input(f"Type 'y' to continue calculating with {answer}, or type 'n' to start a new calculation: \n").lower()
     if keep_going == 'n':
         asnwer = 0
-        num1 = int(input("What's the first number?: \n"))
+        num1 = float(input("What's the first number?: \n"))
         function = input("Pick an operation: \n")
-        num2 = int(input("What's the second number?: \n")) 
+        num2 = float(input("What's the second number?: \n")) 
         answer = operations[function](num1,num2)
         print(f"{num1} {function} {num2} = {answer}") 
         keep_going = input(f"Type 'y' to continue calculating with {answer}, or type 'n' to start a new calculation: \n").lower()
